Removed UnitExample and simplified Button
