--- components/units/UnitExample/UnitExample.spec.tsx
@@ -1,1 +1,0 @@
-export {};
