--- components/units/UnitExample/examples/usage.ts
@@ -1,1 +1,0 @@
-export {};
